Speedo and Odo updates
They work on car side verified with serial port reads, issues
troubleshooting the controller code.  It likes to show random numbers
and occasionally flash over to the right one. Not sure whats up about
that

Bugs with cruise control.  while testing seems to not be very
adjustable. and speed never updated - NOTE: I did not test long with
that feature